Refactor de respository
