+++ src/main/java/com/uis/sisgan/service/CattleService.java
@@ -1,4 +1,4 @@
-package com.uis.sisgan.persistence.service;
+package com.uis.sisgan.service;
 
 import com.uis.sisgan.persistence.crud.CattleCrudRepository;
 import com.uis.sisgan.persistence.entity.Cattle;
